+++ development/next-app/src/pages/_document.tsx
@@ -1,3 +1,4 @@
+import Navbar from '@/components/Navbars/AuthNavbar'
 import { Html, Head, Main, NextScript } from 'next/document'
 
 export default function Document() {
