+++ development/next-app/src/pages/sorteos.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import Navbar from "../components/Navbars/AuthNavbar";
+
+export default function Landing() {
+  return (
+    <>
+      <Navbar />
+    </>
+  );
+}
